Consolidate irListOfStrings helper in IrUtils
Both CompanionGeneration and ObjGeneration had copies of an IR
`listOf<String>(...)` emitter. Fold them into a single helper in
IrUtils.kt so both sites (and any future caller) share one
implementation.
